add boxplot and candlestick stories
